Upgrade build stage to use node:lts-alpine3.13
Resolves: https://github.com/ortelius/ortelius/issues/246

The version of hugo available in the apk repos for our current node builder image is out of date (0.6.x). As a result, we're seeing a bug on ortelius.io around draft: true support for directories.

Updating the node image to  lts-alpine3.13 gives us access to a more recent version of hugo with support for draft directories.