Add .gitignore for hugo and npm generated resources
We have three generated artifact directories:
- Resources
- Public
- node_modules

Due to the fact we generate the site then copy the generated content between docker stages, as opposed to serving from a direct checkout, we can exclude these from the repo.

This commit also removes the public folder which was commited to the repo.